finish moving event loop implementations out of headers
ok

ah so that was the mac build problem
